Fix test compliation error
