+++ ui/src/app/utils/user-apps-utils.tsx
@@ -3,6 +3,7 @@ import {
   AppType,
   CreateAppRequest,
   Disk,
+  DiskType,
   ListAppsResponse,
   UserAppEnvironment,
 } from 'generated/fetch';
@@ -16,6 +17,10 @@ import {
 import { appDisplayPath } from 'app/routing/utils';
 import { leoAppsApi } from 'app/services/notebooks-swagger-fetch-clients';
 import { appsApi } from 'app/services/swagger-fetch-clients';
+import {
+  DEFAULT_AUTODELETE_THRESHOLD_MINUTES,
+  DEFAULT_MACHINE_NAME,
+} from 'app/utils/machines';
 import { userAppsStore } from 'app/utils/stores';
 
 import { MAX_GKE_APP_DISK_SIZE } from './constants';
@@ -35,12 +40,69 @@ export const appTypeToString: Record<AppType, string> = {
 };
 
 export const appMinDiskSize: Record<AppType, number> = {
-  [AppType.CROMWELL]: 50,
-  [AppType.RSTUDIO]: 100,
-  [AppType.SAS]: 150,
+  ['CROMWELL']: 50,
+  ['RSTUDIO']: 100,
+  ['SAS']: 150,
 };
 
 export const appMaxDiskSize = 1000;
+
+// TODO replace with better defaults?
+export const defaultCromwellCreateRequest: CreateAppRequest = {
+  appType: AppType.CROMWELL,
+  kubernetesRuntimeConfig: {
+    numNodes: 1,
+    machineType: DEFAULT_MACHINE_NAME,
+    autoscalingEnabled: false,
+  },
+  persistentDiskRequest: {
+    size: appMinDiskSize.CROMWELL,
+    diskType: DiskType.STANDARD,
+  },
+  autodeleteEnabled: false,
+  // Okay to have value here because when autodeleteEnabled is false, this value is not used.
+  autodeleteThreshold: DEFAULT_AUTODELETE_THRESHOLD_MINUTES,
+};
+
+// TODO replace with better defaults?
+export const defaultRStudioCreateRequest: CreateAppRequest = {
+  appType: AppType.RSTUDIO,
+  kubernetesRuntimeConfig: {
+    numNodes: 1,
+    machineType: DEFAULT_MACHINE_NAME,
+    autoscalingEnabled: false,
+  },
+  persistentDiskRequest: {
+    size: appMinDiskSize[AppType.RSTUDIO],
+    diskType: DiskType.STANDARD,
+  },
+  autodeleteEnabled: false,
+  // Okay to have value here because when autodeleteEnabled is false, this value is not used.
+  autodeleteThreshold: DEFAULT_AUTODELETE_THRESHOLD_MINUTES,
+};
+
+// TODO replace with better defaults?
+export const defaultSASCreateRequest: CreateAppRequest = {
+  appType: AppType.SAS,
+  kubernetesRuntimeConfig: {
+    numNodes: 1,
+    machineType: DEFAULT_MACHINE_NAME,
+    autoscalingEnabled: false,
+  },
+  persistentDiskRequest: {
+    size: appMinDiskSize.SAS,
+    diskType: DiskType.STANDARD,
+  },
+  autodeleteEnabled: false,
+  // Okay to have value here because when autodeleteEnabled is false, this value is not used.
+  autodeleteThreshold: DEFAULT_AUTODELETE_THRESHOLD_MINUTES,
+};
+
+export const defaultAppRequest: Record<AppType, CreateAppRequest> = {
+  [AppType.CROMWELL]: defaultCromwellCreateRequest,
+  [AppType.RSTUDIO]: defaultRStudioCreateRequest,
+  [AppType.SAS]: defaultSASCreateRequest,
+};
 
 const transitionalAppStatuses: Array<AppStatus> = [
   AppStatus.DELETING,
